refactor: migrate exception and IO packages to shared module
Moved `undecided.erp.common.exception` and `undecided.erp.common.io` packages to `undecided.shared.common`. Updated all related package references and resolved minor formatting inconsistencies to enhance modularity and reuse.
