--- backend/src/main/java/undecided/erp/common/exception/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 例外を格納.
- */
-package undecided.erp.common.exception;
