--- backend/src/main/java/undecided/erp/common/io/package-info.java
@@ -1,4 +1,0 @@
-/**
- * 入出力関連のユーティリティ
- */
-package undecided.erp.common.io;
