+++ backend/src/main/java/undecided/erp/greeting/internal/GreetingApi.java
@@ -3,7 +3,7 @@ package undecided.erp.greeting.internal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import undecided.erp.common.exception.BusinessException;
+import undecided.shared.common.exception.BusinessException;
 import undecided.shared.common.message.ResultMessages;
 
 @RestController
