+++ backend/src/main/java/undecided/shared/common/exception/ExceptionCodeResolver.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.exception;
+package undecided.shared.common.exception;
 
 public interface ExceptionCodeResolver {
