+++ backend/src/main/java/undecided/shared/common/exception/ExceptionLogger.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.exception;
+package undecided.shared.common.exception;
 
 import jakarta.annotation.PostConstruct;
 import java.text.MessageFormat;
