+++ backend/src/main/java/undecided/shared/common/exception/NotFoundBusinessException.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.exception;
+package undecided.shared.common.exception;
 
 import undecided.shared.common.message.ResultMessages;
