+++ backend/src/main/java/undecided/shared/common/exception/SystemException.java
@@ -1,31 +1,26 @@
-package undecided.erp.common.exception;
+package undecided.shared.common.exception;
 
 /**
  * このクラスは RuntimeException クラスのサブクラスであり、ExceptionCodeProvider インターフェースを実装する システム例外を表します。
- * <p>
- * SystemException クラスには3つのコンストラクタがあり、それぞれ例外コード、表示されるメッセージ、および 例外の根本的な原因を指定できます。code パラメータは
+ *
+ * <p>SystemException クラスには3つのコンストラクタがあり、それぞれ例外コード、表示されるメッセージ、および 例外の根本的な原因を指定できます。code パラメータは
  * ExceptionCodeProvider 型であり、 ExceptionCodeProvider インターフェースの実装である必要があります。message パラメータは例外がスロー
  * されたときに表示されるメッセージを表す文字列です。cause パラメータは例外の根本的な原因を表す Throwable オブジェクトです。
- * <p>
- * SystemException クラスは、例外に関連付けられた例外コードを返す ExceptionCodeProvider インターフェース の getCode()
+ *
+ * <p>SystemException クラスは、例外に関連付けられた例外コードを返す ExceptionCodeProvider インターフェース の getCode()
  * メソッドをオーバーライドしています。
  */
-
-public class SystemException extends RuntimeException implements
-    ExceptionCodeProvider {
+public class SystemException extends RuntimeException implements ExceptionCodeProvider {
 
   private static final long serialVersionUID = 1L;
 
-  /**
-   * exception code.
-   */
+  /** exception code. */
   private final String code;
 
   /**
    * コンストラクタ
-   * <p>
-   * {@link ExceptionCodeProvider}、表示されるメッセージ、および例外の根本原因を指定できます。
-   * </p>
+   *
+   * <p>{@link ExceptionCodeProvider}、表示されるメッセージ、および例外の根本原因を指定できます。
    *
    * @param code 例外コード {@link ExceptionCodeProvider}
    * @param message 表示されるメッセージ
@@ -38,9 +33,8 @@ public class SystemException extends RuntimeException implements
 
   /**
    * コンストラクタ
-   * <p>
-   * {@link ExceptionCodeProvider} と表示されるメッセージを指定できます。
-   * </p>
+   *
+   * <p>{@link ExceptionCodeProvider} と表示されるメッセージを指定できます。
    *
    * @param code 例外コード {@link ExceptionCodeProvider}
    * @param message 表示されるメッセージ
@@ -52,9 +46,8 @@ public class SystemException extends RuntimeException implements
 
   /**
    * コンストラクタ
-   * <p>
-   * {@link ExceptionCodeProvider} と例外の原因を指定できます。
-   * </p>
+   *
+   * <p>{@link ExceptionCodeProvider} と例外の原因を指定できます。
    *
    * @param code 例外コード {@link ExceptionCodeProvider}
    * @param cause 例外の原因となる要素
@@ -64,7 +57,6 @@ public class SystemException extends RuntimeException implements
     this.code = code;
   }
 
-
   /**
    * Retrieves the exception code associated with the system exception.
    *
@@ -73,7 +65,5 @@ public class SystemException extends RuntimeException implements
   @Override
   public String getCode() {
     return code;
-
   }
-
 }
