+++ backend/src/main/java/undecided/shared/common/exception/package-info.java
@@ -1,0 +1,2 @@
+/** 例外を格納. */
+package undecided.shared.common.exception;
