+++ backend/src/main/java/undecided/shared/common/io/TextFiles.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.io;
+package undecided.shared.common.io;
 
 import java.io.InputStream;
 import lombok.experimental.UtilityClass;
@@ -8,8 +8,8 @@ public class TextFiles {
 
   /**
    * UTF-8のバイトオーダーマーク (BOM) をバイト配列として表現します。
-   * <p>
-   * この定数は、UTF-8でエンコードされたテキストストリームの先頭に 付加される3バイトのシーケンスを含みます。このシーケンスはテキスト
+   *
+   * <p>この定数は、UTF-8でエンコードされたテキストストリームの先頭に 付加される3バイトのシーケンスを含みます。このシーケンスはテキスト
    * ファイルのエンコーディングを示すためのマーカーとして使用されます。 ただし、RFC 3629によるとUTF-8のファイルでは使用は任意であり、
    * 推奨されていません。具体的なバイト値は以下の通りです: - 0xEF - 0xBB - 0xBF
    */
@@ -18,5 +18,4 @@ public class TextFiles {
   public static boolean existsUtf8Bom(InputStream inputStream) {
     return inputStream.markSupported();
   }
-
 }
