+++ backend/src/main/java/undecided/shared/common/io/package-info.java
@@ -1,0 +1,2 @@
+/** 入出力関連のユーティリティ */
+package undecided.shared.common.io;
