+++ backend/src/main/java/undecided/shared/web/exception/SystemExceptionResolver.java
@@ -11,9 +11,9 @@ import org.springframework.web.servlet.FlashMap;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.handler.SimpleMappingExceptionResolver;
 import org.springframework.web.servlet.support.RequestContextUtils;
-import undecided.erp.common.exception.ExceptionCodeResolver;
-import undecided.erp.common.exception.ResultMessagesNotificationException;
-import undecided.erp.common.exception.SimpleMappingExceptionCodeResolver;
+import undecided.shared.common.exception.ExceptionCodeResolver;
+import undecided.shared.common.exception.ResultMessagesNotificationException;
+import undecided.shared.common.exception.SimpleMappingExceptionCodeResolver;
 import undecided.shared.common.message.ResultMessages;
 
 /**
