+++ backend/src/test/java/undecided/shared/common/exception/SystemExceptionTest.java
@@ -1,10 +1,10 @@
-package undecided.erp.common.exception;
+package undecided.shared.common.exception;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("SystemException のテスト")
 class SystemExceptionTest {
